% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisherEl22Part2.R
\name{fisherEl22Part2}
\alias{fisherEl22Part2}
\title{Part 2 of element [2, 2] for Fisher Information matrix}
\usage{
fisherEl22Part2(nvData, nTheta1, nTheta2, lDensityExpr)
}
\arguments{
\item{nvData}{The vector of data.}

\item{nTheta1}{The first parameter.}

\item{nTheta2}{The second parameter.}

\item{lDensityExpr}{List of symbolic expressions of density, cumulative and derivatives.}
}
\description{
Function computing part 2 of element [2, 2] for Fisher Information matrix computation. The Fisher Information matrix is
splitted in the four elements ([1, 1], [1, 2], [2, 1], [2, 2]). Each element is split in part 1 and part 2
}
\keyword{internal}

