% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matFisherComputation.R
\name{matFisherComputation}
\alias{matFisherComputation}
\title{Fisher information matrix}
\usage{
matFisherComputation(nTheta1, nTheta2, lDensityExpr)
}
\arguments{
\item{nTheta1}{First parameter.}

\item{nTheta2}{Second parameter.}

\item{lDensityExpr}{List of symbolic expressions of density, cumulative and derivatives.}
}
\value{
The Fisher information matrix.
}
\description{
Function calculating the Fisher information matrix.
}
