% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.R
\name{MLE}
\alias{MLE}
\title{Numerical Maximum Likelihood Estimator}
\usage{
MLE(nvData, strDistribution, lDensityExpr)
}
\arguments{
\item{nvData}{The vector of the data.}

\item{strDistribution}{The distribution name.}

\item{lDensityExpr}{The distribution expression,}
}
\value{
A list with distribution name, distribution parameters, value of the objective function
corresponding to the parameters, additional information returned by the optimizer, convergence of
the algorithm.
}
\description{
The parameters Maximum Likelihood Estimation is obtained by numerical optimization.
}
