% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OBRECheckTolParameters.R
\name{OBRECheckTolParameters}
\alias{OBRECheckTolParameters}
\title{Check if OBRE matrix A and vector a are final.}
\usage{
OBRECheckTolParameters(matANew, matAOld, nvANew, nvAOld, nRelTol, nAbsTol)
}
\arguments{
\item{matANew}{Matrix A at the current iteration.}

\item{matAOld}{Matrix A at the past iteration.}

\item{nvANew}{Vector a at the current iteration.}

\item{nvAOld}{Vector a at the past iteration.}

\item{nRelTol}{Relative tolerance.}

\item{nAbsTol}{Absolute tolerance.}
}
\value{
A flag indicating if condition on matrix A and vector a are both satisfied.
}
\description{
The function compute the relative distance from the past to the current iteration of matrix A, with
respect to the relative tolerance if at the current iteration matrix A is not null. Otherwise the absolute
error is checked. Then the vector a is checked in the same way.
}
