% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OBREmatMArgumentB.R
\name{OBREmatMArgumentB}
\alias{OBREmatMArgumentB}
\title{Argument B for OBRE matrix M integrals.}
\usage{
OBREmatMArgumentB(
  nvData,
  nTheta1,
  nTheta2,
  lDensityExpr,
  nCParOBRE,
  matA,
  nvA,
  nK
)
}
\arguments{
\item{nvData}{The vector of data.}

\item{nTheta1}{The first parameter.}

\item{nTheta2}{The second parameter.}

\item{lDensityExpr}{List of symbolic expressions of density, cumulative and derivatives.}

\item{nCParOBRE}{OBRE c parameter.}

\item{matA}{Matrix A.}

\item{nvA}{Vector a.}

\item{nK}{Exponent which differentiate M_1 from M_2.}
}
\description{
Function computing argument B for OBRE matrix M integrals.
}
