% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OBREMatMComputation.R
\name{OBREMatMComputation}
\alias{OBREMatMComputation}
\title{Function computing the OBRE matrix M.}
\usage{
OBREMatMComputation(nvData, nTheta1, nTheta2, lDensityExpr, nCParOBRE, matA,
  nvA, nK)
}
\arguments{
\item{nvData}{The vector of data.}

\item{nTheta1}{The first parameter.}

\item{nTheta2}{The second parameter.}

\item{lDensityExpr}{List of symbolic expressions of density, cumulative and derivatives.}

\item{nCParOBRE}{OBRE c parameter.}

\item{matA}{Matrix A.}

\item{nvA}{Vector a.}

\item{nK}{Exponent which differentiate M_1 from M_2.}
}
\value{
OBRE M matrix (M_1 if nK = 1; M_2 if nK = 2).
}
\description{
The function evaluates integrals used to compute the M_1 and M_2 OBRE matrices. Element (1,1) uses
argument (A,B,F); element (1,2) uses argument (B,D,E,F); elements (2,2) uses arguments (C,D,F).
}
\keyword{internal}

