% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OBREMatVMatMEl21.R
\name{OBREMatVMatMEl21}
\alias{OBREMatVMatMEl21}
\title{Element [2, 1] of matrix M.}
\usage{
OBREMatVMatMEl21(nvData, nTheta1, nTheta2, lDensityExpr, nCParOBRE, matA, nvA)
}
\arguments{
\item{nvData}{The vector of data.}

\item{nTheta1}{The first parameter.}

\item{nTheta2}{The second parameter.}

\item{lDensityExpr}{List of symbolic expressions of density, cumulative and derivatives.}

\item{nCParOBRE}{OBRE c parameter.}

\item{matA}{Matrix A.}

\item{nvA}{Vector a.}
}
\description{
Function computing element [2, 1] of matrix M, for the computation of asymptotic covariance matrix V.
}
\keyword{internal}

