% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OBREMatVMatQEl11.R
\name{OBREMatVMatQEl11}
\alias{OBREMatVMatQEl11}
\title{Element [1, 1] of matrix Q.}
\usage{
OBREMatVMatQEl11(nvData, nTheta1, nTheta2, lDensityExpr, nCParOBRE, matA, nvA)
}
\arguments{
\item{nvData}{The vector of data.}

\item{nTheta1}{The first parameter.}

\item{nTheta2}{The second parameter.}

\item{lDensityExpr}{List of symbolic expressions of density, cumulative and derivatives.}

\item{nCParOBRE}{OBRE c parameter.}

\item{matA}{Matrix A.}

\item{nvA}{Vector a.}
}
\description{
Function computing argument element [1, 1] of matrix Q of asymptotic covariance matrix V.
}
