% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OBREnvAComputation.R
\name{OBREnvAComputation}
\alias{OBREnvAComputation}
\title{OBRE vector a.}
\usage{
OBREnvAComputation(
  nvData,
  nTheta1,
  nTheta2,
  lDensityExpr,
  nCParOBRE,
  matA,
  nvA
)
}
\arguments{
\item{nvData}{The vector of data.}

\item{nTheta1}{The first parameter.}

\item{nTheta2}{The second parameter.}

\item{lDensityExpr}{The list of symbolic expressions of density, cumulative and derivatives.}

\item{nCParOBRE}{OBRE c parameter.}

\item{matA}{OBRE matrix A.}

\item{nvA}{OBRE vector a.}
}
\value{
The OBRE a vector.
}
\description{
The function evaluates integrals used to compute the components of OBRE a vector.
}
