% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OBREnvADen.R
\name{OBREnvADen}
\alias{OBREnvADen}
\title{Denominator for nvA}
\usage{
OBREnvADen(nvData, nTheta1, nTheta2, lDensityExpr, nCParOBRE, matA, nvA)
}
\arguments{
\item{nvData}{The vector of data.}

\item{nTheta1}{The first parameter.}

\item{nTheta2}{The second parameter.}

\item{lDensityExpr}{List of symbolic expressions of density, cumulative and derivatives.}

\item{nCParOBRE}{OBRE c parameter.}

\item{matA}{Matrix A.}

\item{nvA}{Vector a.}
}
\description{
Function computing denominator for OBRE numeric vector nvA evaluation.
}
