% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OBREnvANum1.R
\name{OBREnvANum1}
\alias{OBREnvANum1}
\title{First part numerator for nvA}
\usage{
OBREnvANum1(nvData, nTheta1, nTheta2, lDensityExpr, nCParOBRE, matA, nvA)
}
\arguments{
\item{nvData}{The vector of data.}

\item{nTheta1}{The first parameter.}

\item{nTheta2}{The second parameter.}

\item{lDensityExpr}{List of symbolic expressions of density, cumulative and derivatives.}

\item{nCParOBRE}{OBRE c parameter.}

\item{matA}{Matrix A.}

\item{nvA}{Vector a.}
}
\description{
Function computing first part numerator for OBRE numeric vector nvA evaluation.
}
\keyword{internal}

