% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OBREWeightsFun.R
\name{OBREWeightsFun}
\alias{OBREWeightsFun}
\title{OBRE weights.}
\usage{
OBREWeightsFun(nvData, nTheta1, nTheta2, lDensityExpr, nCParOBRE, matA, nvA)
}
\arguments{
\item{nvData}{The vector of data.}

\item{nTheta1}{The first parameter.}

\item{nTheta2}{The second parameter.}

\item{lDensityExpr}{The list of symbolic expressions of density, cumulative and derivatives.}

\item{nCParOBRE}{OBRE c parameter.}

\item{matA}{OBRE matrix A.}

\item{nvA}{OBRE vector a.}
}
\value{
A numeric vector containing OBRE weights.
}
\description{
Function for computing OBRE weights. The function computes the score function for both parameters and
build the score matrix. The score matrix is then modified using OBRE parameters A matrix and a vector and
an euclidean norm is derived. The weights are finally found as the minimum between the normalized nCParOBRE
and 1.
}
\keyword{internal}

