% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreComponent.R
\name{scoreComponent}
\alias{scoreComponent}
\title{First component of the score function.}
\usage{
scoreComponent(nvData, nTheta1, nTheta2, lDensityExpr, nParIndex)
}
\arguments{
\item{nvData}{The vector of data.}

\item{nTheta1}{The first parameter.}

\item{nTheta2}{The second parameter.}

\item{lDensityExpr}{The list of symbolic expressions of density, cumulative and derivatives.}

\item{nParIndex}{Which component parameter needs to be calculated.}
}
\value{
The first component of the score function.
}
\description{
The function evaluates the formula used to compute the first component of the score function. The
missing elements are imputed with 0.
}
